Sharing an object test in Python
This commit moves the previous "test" folder to "examples" instead and replaces the test folder with a place for bats/bash tests. This commit also adds a feature test that proves sharing an object can be done by creating a manual container, and proves that Python DI construction is also currently supported.

Closes #1 (with feature test see #5)